requirements.txt
argon2-cffi==20.1.0
asgiref==3.2.10
astroid==2.3.2
async-generator==1.10
attrs==20.2.0
autocommand==2.2.1
autopep8==1.5.4
backcall==0.2.0
bleach==3.2.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
colorama==0.4.1
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
Django==3.1
django-allauth==0.42.0
django-cors-headers==3.4.0
django-crispy-forms==1.11.2
django-money==1.1
django-rest-auth==0.9.5
djangorestframework==3.11.0
entrypoints==0.3
idna==2.10
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jedi==0.17.2
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.3.3
mccabe==0.6.1
mistune==0.8.4
mysqlclient==2.0.1
nbclient==0.5.0
nbconvert==6.0.7
nbformat==5.0.7
nest-asyncio==1.4.1
networkx==2.5
notebook==6.1.4
numpy==1.19.4
oauthlib==3.1.0
packaging==20.4
pandocfilters==1.4.2
parso==0.7.1
path==16.3.0
pep8==1.7.1
pickleshare==0.7.5
Pillow==7.2.0
pip-run==8.7.2
prometheus-client==0.8.0
prompt-toolkit==3.0.8
py-moneyed==0.8.0
pycodestyle==2.6.0
pycparser==2.20
Pygments==2.7.1
pylint==2.4.3
pyparsing==2.4.7
pyrsistent==0.17.3
python-bitcoinrpc==1.0
python-dateutil==2.8.1
python-dotenv==0.19.2
python3-openid==3.2.0
pytz==2020.1
pywin32==228
pywinpty==0.5.7
pyzmq==19.0.2
qtconsole==4.7.7
QtPy==1.9.0
requests==2.24.0
requests-oauthlib==1.3.0
Send2Trash==1.5.0
six==1.12.0
sqlparse==0.3.1
terminado==0.9.1
testpath==0.4.4
toml==0.10.1
tornado==6.0.4
traitlets==5.0.4
urllib3==1.25.10
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.11.2
